chore: split migrations based on conduit version
