+++ scripts/migrateV10_V15/migrationFiles/index.js
@@ -1,4 +1,5 @@
-export * from './authenticationMigrations'
-export * from './database/databaseMigrations'
-export * from './database/customEndpointsMigrations'
-export * from './storageMigrations'
+export * from './migrateV10_V11'
+export * from './migrateV11_V12'
+export * from './migrateV13_V14'
+export * from './migrateV14_V15'
+
