+++ scripts/migrateV10_V15/migrationFiles/migrateV11_V12.js
@@ -1,25 +1,11 @@
-const db = require('../../mongoConnection');
+const db = require('../mongoConnection');
 const { isNil, merge } = require('lodash');
 
 
-const migrateV11_V15 = async () => {
-  const documents = db.collection('declaredschemas');
-  const schemas = await documents.find({ 'modelOptions.conduit': { $exists: false } }).toArray();
-  for (const schema of schemas) {
-    let newModelOptions = { conduit: {} };
-    try {
-      newModelOptions = { ...newModelOptions, ...JSON.parse((schema.modelOptions)) };
-    } catch {
-      newModelOptions = { ...newModelOptions, ...schema.modelOptions };
-    }
-    await documents.updateMany(schema._id, { $set: { modelOptions: newModelOptions } }).catch((err) => {
-      console.log(err);
-    });
-  }
-};
+
 
 const migrateV12_V15_cmsOwners = async () => {
-  const documents = db.collection('declaredschemas');
+  const documents = db.collection('_declaredschemas');
   const schemas = await documents.find({ ownerModule: 'cms' }).toArray();
   if (schemas.filter((schema) => schema.name !== 'schemadefinitions').length === 0)
     return;
@@ -38,7 +24,7 @@ const migrateV12_V15_cmsOwners = async () => {
 };
 
 const migrateV12_V15_customEndpoints = async () => {
-  const documents = db.collection('declaredschemas');
+  const documents = db.collection('_declaredschemas');
   const declaredSchemas = await documents.find({ 'modelOptions.conduit': { $exists: false } }).toArray();
   for (const declaredSchema of declaredSchemas) {
     let newModelOptions = { conduit: {} };
@@ -47,14 +33,14 @@ const migrateV12_V15_customEndpoints = async () => {
     } catch {
       newModelOptions = { ...newModelOptions, ...declaredSchema.modelOptions };
     }
-    await documents.findOneAndUpdate(declaredSchema._id, { modelOptions: newModelOptions }).catch((err) => {
+    await documents.updateOne({ _id: declaredSchema._id }, { modelOptions: newModelOptions }).catch((err) => {
       console.log(err);
     });
   }
 };
 
 const migrateV12_V15_modelOptions = async () => {
-  const documents = db.collection('declaredschemas');
+  const documents = db.collection('_declaredschemas');
   const cmsSchemas = await documents.find({
     $or: [
       { 'modelOptions.conduit.cms.crudOperations': true },
@@ -85,7 +71,7 @@ const migrateV12_V15_modelOptions = async () => {
       },
     };
     const id = schema._id.toString();
-    await documents.updateMany({ id }, {
+    await documents.updateMany({ _id : id}, {
       $set: {
         modelOptions: {
           conduit: { cms },
@@ -96,7 +82,7 @@ const migrateV12_V15_modelOptions = async () => {
 };
 
 const migrateV12_V15_schemaDefinitions = async () => {
-  const documents = db.collection('declaredschemas');
+  const documents = db.collection('_declaredschemas');
   const schemas = await documents.find().toArray();
   if (schemas.filter((schema) => schema.name === 'schemadefinitions').length === 0)
     return;
@@ -107,7 +93,7 @@ const migrateV12_V15_schemaDefinitions = async () => {
   // Migrate SchemaDefinitions to DeclaredSchemas
   if (!isNil(schemaDefinitions)) {
     for (const schema of schemaDefinitions) {
-      const declaredSchema = await db.collection('declaredschemas')
+      const declaredSchema = await db.collection('_declaredschemas')
         .findOne({ name: schema.name });
       let modelOptions = {
         conduit: {
@@ -147,49 +133,28 @@ const migrateV12_V15_schemaDefinitions = async () => {
   }
 };
 
-const migrateV14_V15 = async () => {
-  const documents = db.collection('declaredschemas');
-  await documents.updateMany({ name: 'Client' }, { $set: { ownerModule: 'router' } });
+const migrateV12_V15 = async () => {
+  const documents = db.collection('configs');
+  const authConfig = await documents.findOne({"moduleConfigs.authentication": {$exists: true}});
+  if (authConfig.local['verificationRequired']) {
+    authConfig.local.verification = {
+      required: authConfig.local['verificationRequired'],
+      send_email: authConfig.local['sendVerificationEmail'],
+      redirect_uri: authConfig.local['verification_redirect_uri'],
+    };
+    delete authConfig.local['verificationRequired'];
+    delete authConfig.local['sendVerificationEmail'];
+    delete authConfig.local['verification_redirect_uri'];
+    await documents.findOneAndUpdate(authConfig._id, authConfig);
+  }
 };
 
-const migrateV15 = async () => {
-  const collection = db.collection('declaredschemas');
-  const declaredSchemas = await collection.find({ $and: [
-      { name: { $in: adapter.systemSchemas } },
-      { 'modelOptions.conduit.cms': { $exists: true } },
-    ]}).toArray();
-  /*
-  *   const declaredSchemas = adapter.getSchemaModel('_DeclaredSchema').model;
-  const affectedSchemas: IDeclaredSchema[] = await declaredSchemas.findMany({
-    $and: [
-      { name: { $in: adapter.systemSchemas } },
-      { 'modelOptions.conduit.cms': { $exists: true } },
-    ],
-  });
-
-  if (affectedSchemas.length > 0) {
-    for (const schema of affectedSchemas) {
-      const conduit = schema.modelOptions.conduit;
-      delete conduit!.cms;
-      await declaredSchemas.findByIdAndUpdate(schema._id, {
-        modelOptions: {
-          conduit,
-        },
-      });
-    }
-    const customEndpoints = adapter.getSchemaModel('CustomEndpoints').model;
-    const affectedSchemaNames = affectedSchemas.map(s => s.name);
-    await customEndpoints.deleteMany({ selectedSchema: { $in: affectedSchemaNames } });
-  }*/
-};
-const migrate = async () => {
-  await migrateV11_V15();
-  await migrateV12_V15_cmsOwners();
-  await migrateV12_V15_customEndpoints();
-  await migrateV12_V15_modelOptions();
+const migrateV11_V12 = async () => {
   await migrateV12_V15_schemaDefinitions();
-  await migrateV14_V15();
-  await migrateV15();
-};
+  await migrateV12_V15_modelOptions();
+  await migrateV12_V15_customEndpoints();
+  await migrateV12_V15_cmsOwners();
+  await migrateV12_V15();
+}
 
-migrate().then(r => console.log(r)).catch(e => console.log(e));
+module.exports = migrateV11_V12;
